# AS::Callbacks: specialize call templates for better performance