By doing so we avoid a lot of extra work.

```
            baseline:  4092604.4 i/s
          opt-method:   693204.7 i/s - 5.90x  (± 0.00) slower
              method:   614761.0 i/s - 6.66x  (± 0.00) slower
```

Baseline is calling `run_callbacks` with no callbacks registered.

Full benchmark: https://gist.github.com/casperisfine/837a7a665c6b232dadcf980d73694748

The same pattern could be applied to the other branches, but I focused
on method calls as they're the most common.

### Potential followups

- We could apply the same pattern to the various `instance_exec` forms.
- StackProf shows that my benchmark spend about `67.0%` if it's time in gc sweep, so we could try to reduce allocations too.

cc @jhawthorn @rafaelfranca