@casperisfine 2021-11-24
So I went ahead and applied the same transformation to the other argument types.

I also did a different benchmark:

```ruby
require 'benchmark/ips'
require 'active_support/all'

class Base
  include ActiveSupport::Callbacks
  define_callbacks :save

  def save
    run_callbacks :save do
    end
  end
end

class Base2 < Base
  set_callback :save, :before do |record|
  end
  set_callback :save, :before, :my_callback

  def my_callback
  end
end

class Record < Base2
  set_callback :save, :before do |record|
  end
  set_callback :save, :before, :my_callback2

  def my_callback2
  end
end

BASELINE = Base.new
RECORD = Record.new

Benchmark.ips do |x|
  x.report('baseline') { BASELINE.save }
  x.report('callbacks') { RECORD.save }

  x.compare!
end
```

On `main`:
```
Warming up --------------------------------------
            baseline   424.176k i/100ms
           callbacks    21.157k i/100ms
Calculating -------------------------------------
            baseline      4.235M (± 0.9%) i/s -     21.209M in   5.008156s
           callbacks    210.047k (± 1.3%) i/s -      1.058M in   5.037139s

Comparison:
            baseline:  4235234.0 i/s
           callbacks:   210047.2 i/s - 20.16x  (± 0.00) slower
```

this branch:
```
Warming up --------------------------------------
            baseline   422.975k i/100ms
           callbacks    25.810k i/100ms
Calculating -------------------------------------
            baseline      4.250M (± 1.7%) i/s -     21.572M in   5.076946s
           callbacks    262.126k (± 1.0%) i/s -      1.316M in   5.022193s

Comparison:
            baseline:  4250279.4 i/s
           callbacks:   262126.5 i/s - 16.21x  (± 0.00) slower
```

So the overhead is down from 20x to 16x which is more or less a 20% improvement, but there's lots of extra code being benchmarked.